Django_assignment/membership_project/pyproject.toml
[tool.poetry]
name = "membership-project"
version = "0.1.0"
description = ""
authors = ["hyo <umdoongs@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
django = "^5.1.4"
djangorestframework = "^3.15.2"
djangorestframework-simplejwt = "^5.3.1"
django-environ = "^0.11.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
